Extract fshash.Metadata to its own file.
