+++ lib/fshash/metadata_test.go
